# api-specs/schack-se-api-docs.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://member.schack.se/",
      "description": "Default Server URL"
    }
  ],
  "paths": {
...
    "/public/api/v1/tournamentresults/team/table/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta turneringstabell (klubblagturnering) för en grupp i en turnering",
        "operationId": "getTeamResultTable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamTournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/team/roundresults/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten (klubblagturnering) för en grupp i en turnering",
        "operationId": "getTeamRoundResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/team/roundresults/id/{id}/memberid/{memberid}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten för en spelare i en turnering",
        "operationId": "getTeamRoundResults_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "memberid",
            "in": "path",
            "description": "id för medlemmen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/table/memberid/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta individuella turneringsresultat/anmälningar för en medlem",
        "operationId": "getMemberTableResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för medlemmen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/table/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta turneringstabell (endast individuell än så länge) för en grupp i en turnering",
        "operationId": "getResultTable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/roundresults/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten (endast individuell än så länge) för en grupp i en turnering",
        "operationId": "getRoundResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/ratinglist/district/{id}/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för ett distrikt, givet datum, rankingkategori och medlemskategori",
        "operationId": "getDistrictRatingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för distriktet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/ratinglist/club/{id}/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för en klubb, givet datum, rankingkategori och medlemskategori",
        "operationId": "getClubRatingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för klubben",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MemberFIDERatingDTO": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "k": {
            "type": "integer",
            "format": "int32"
          },
          "rapidRating": {
            "type": "integer",
            "format": "int32"
          },
          "rapidk": {
            "type": "integer",
            "format": "int32"
          },
          "blitzRating": {
            "type": "integer",
            "format": "int32"
          },
          "blitzK": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MemberLASKRatingDTO": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PlayerInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthdate": {
            "type": "string",
            "format": "date-time"
          },
          "sex": {
            "type": "integer",
            "format": "int32"
          },
          "fideid": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "club": {
            "type": "string"
          },
          "clubId": {
            "type": "integer",
            "format": "int32"
          },
          "elo": {
            "$ref": "#/components/schemas/MemberFIDERatingDTO"
          },
          "lask": {
            "$ref": "#/components/schemas/MemberLASKRatingDTO"
          }
        }
      },
...
      "ClubDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "zipcode": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string"
          },
          "startdate": {
            "type": "string",
            "format": "date-time"
          },
          "startSeason": {
            "type": "string",
            "format": "date-time"
          },
          "endSeason": {
            "type": "string",
            "format": "date-time"
          },
          "phonenr": {
            "type": "string"
          },
          "postgiro": {
            "type": "string"
          },
          "alliansclub": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "orgnumber": {
            "type": "string"
          },
          "districts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistrictMembershipDTO"
            }
          },
          "regYear": {
            "$ref": "#/components/schemas/RegistrationYear"
          },
          "url": {
            "type": "string"
          },
          "vbdescr": {
            "type": "string"
          },
          "schoolClub": {
            "type": "integer",
            "format": "int32"
          },
          "schoolName": {
            "type": "string"
          },
          "schoolid": {
            "type": "integer",
            "format": "int32"
          },
          "lan": {
            "type": "integer",
            "format": "int32"
          },
          "noEconomy": {
            "type": "integer",
            "format": "int32"
          },
          "specialType": {
            "type": "integer",
            "format": "int32"
          },
          "yes2chessreason": {
            "type": "boolean"
          },
          "schack4anreason": {
            "type": "boolean"
          },
          "ovrigtreason": {
            "type": "boolean"
          },
          "hasRatingPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "sisuid": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DistrictMembershipDTO": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "districtid": {
            "type": "integer",
            "format": "int32"
          },
          "clubid": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "integer",
            "format": "int32"
          },
          "benefit": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RegistrationYear": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TeamTournamentEndResultDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number",
            "format": "float"
          },
          "secPoints": {
            "type": "number",
            "format": "double"
          },
          "place": {
            "type": "integer",
            "format": "int32"
          },
          "contenderId": {
            "type": "integer",
            "format": "int32"
          },
          "teamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "wonGames": {
            "type": "integer",
            "format": "int32"
          },
          "drawGames": {
            "type": "integer",
            "format": "int32"
          },
          "lostGames": {
            "type": "integer",
            "format": "int32"
          },
          "club": {
            "$ref": "#/components/schemas/ClubDTO"
          }
        }
      },
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentResultID": {
            "type": "integer",
            "format": "int32"
          },
          "tableNr": {
            "type": "integer",
            "format": "int32"
          },
          "whiteId": {
            "type": "integer",
            "format": "int32"
          },
          "blackId": {
            "type": "integer",
            "format": "int32"
          },
          "result": {
            "type": "integer",
            "format": "int32"
          },
          "pgn": {
            "type": "string"
          },
          "groupiD": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TournamentRoundResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupdId": {
            "type": "integer",
            "format": "int32"
          },
          "roundNr": {
            "type": "integer",
            "format": "int32"
          },
          "board": {
            "type": "integer",
            "format": "int32"
          },
          "homeId": {
            "type": "integer",
            "format": "int32"
          },
          "homeTeamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "awayId": {
            "type": "integer",
            "format": "int32"
          },
          "awayTeamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "homeResult": {
            "type": "number",
            "format": "float"
          },
          "awayResult": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "finalized": {
            "type": "boolean"
          },
          "publisher": {
            "type": "integer",
            "format": "int32"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "publishedNote": {
            "type": "string"
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          }
        }
      },
      "TournamentEndResultDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number",
            "format": "float"
          },
          "secPoints": {
            "type": "number",
            "format": "double"
          },
          "place": {
            "type": "integer",
            "format": "int32"
          },
          "contenderId": {
            "type": "integer",
            "format": "int32"
          },
          "teamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "wonGames": {
            "type": "integer",
            "format": "int32"
          },
          "drawGames": {
            "type": "integer",
            "format": "int32"
          },
          "lostGames": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "playerInfo": {
            "$ref": "#/components/schemas/PlayerInfoDto"
          }
        }
      },
...
    }
  }
}